Replace font use with svg for tag line
Noticed that even with the webfont, still don't have precise control over the
render if the user possible has the font installed on their system. In my case,
after installing the font, the system font overrode the webfont, and didn't have
the correct font weights, so bad render :/.

Rendering to an svg in inkscape provides better assurance. I also tweaked the
of, to read more clearly.